Add unit tests
This uses the showbeam example function to generate Matplotlib images,
which can also be compared to baseline images (generated by setting
the GENERATE_IMAGES flag to True momentarily).

This currently uses Matplotlib's internal pytest tools, which has the
advantage that it is self-contained and also supports image filenames
generated by fixtures. The downside is extra machinery to generate the
images and it does not work on Python 2 (baseline_images may not be
None in matplotlib 2.2.5).

This commit is mostly for documentary purposes. The next iteration will
consider pytest-mpl instead, which is the preferred way to access
Matplotlib testing tools according to their developers (and it works on
Python 2).