Move known models to module constants
This is more Pythonic. Open the tables with np.loadtxt, which is
clearer and also makes future config files an easy upgrade. For the
sake of Python 2, store the model tables in Unicode literal strings to
allow io.StringIO on both Python versions.

Raise ValueError if the model name is unknown instead of just printing
an error.